feeltrack_backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.1.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pyjwt==2.8.0
